backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
PyGithub==2.1.1
beautifulsoup4==4.12.2
markdown==3.5.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4

# Google Gemini API
google-generativeai==0.3.2

# RAG Dependencies
chromadb==0.4.22
langchain==0.1.0
langchain-google-genai==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# NLP Dependencies
spacy==3.7.2
nltk==3.8.1
transformers==4.36.2

# Code Analysis
tree-sitter==0.20.4
pygments==2.17.2
astroid==3.0.2
radon==6.0.1

# Documentation Generation
pydoc-markdown==4.8.2
mkdocs==1.5.3
mkdocs-material==9.5.3
reportlab==4.0.8

# Code Execution
docker==7.0.0
black==23.12.1

# Visualization
networkx==3.2.1
matplotlib==3.8.2

# WebSocket support
websockets==12.0 